+++ src/TSID/include/BipedalLocomotion/TSID/SE3Task.h
@@ -80,7 +80,7 @@ public:
      * |            `kd_angular`            | `double` |                         Gain of the angular velocity controller                        |    Yes    |
      * @return True in case of success, false otherwise.
      */
-    bool initialize(std::weak_ptr<ParametersHandler::IParametersHandler> paramHandler);
+    bool initialize(std::weak_ptr<const ParametersHandler::IParametersHandler> paramHandler);
 
     /**
      * Set the kinDynComputations object.