+++ src/TSID/src/SO3Task.cpp
@@ -67,7 +67,7 @@ bool SO3Task::setKinDyn(std::shared_ptr<iDynTree::KinDynComputations> kinDyn)
     return true;
 }
 
-bool SO3Task::initialize(std::weak_ptr<ParametersHandler::IParametersHandler> paramHandler)
+bool SO3Task::initialize(std::weak_ptr<const ParametersHandler::IParametersHandler> paramHandler)
 {
     constexpr std::string_view errorPrefix = "[SO3Task::initialize] ";